pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "Powerful, automated research and analysis across your own document collections."
authors = [
  { name = "Statistisches Amt Kanton Zürich, Team Data", email = "datashop@statistik.zh.ch" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "pandarallel>=1.6.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.5",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0",
    "sentence-transformers>=4.1.0",
    "spacy>=3.8.7",
    "tqdm>=4.67.1",
    "de-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-3.8.0/de_core_news_lg-3.8.0-py3-none-any.whl",
    "watchdog>=6.0.0",
    "notebook>=7.4.3",
    "pyarrow>=20.0.0",
    "weaviate-client>=4.15.4",
    "tiktoken>=0.9.0",
    "streamlit>=1.46.1",
    "openai>=1.95.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-clarity>=1.0.1",
    "ruff>=0.11.8",
]
